refactor: rearrange Icons folder
