+++ src/components/Icons/SVGComponents/AppIcon.tsx
