+++ src/components/Icons/SVGComponents/Close.tsx
