+++ src/components/Icons/SVGComponents/DeleteIcon.tsx
