+++ src/components/Icons/SVGComponents/DoubleArrow.tsx
