+++ src/components/Icons/SVGComponents/Pause.tsx
