+++ src/components/Icons/SVGComponents/SettingsIcon.tsx
