+++ src/components/Icons/SVGComponents/Speaker.tsx
